backend/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cryptography==40.0.2
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastapi==0.83.0
h11==0.13.0
h5py==3.1.0
idna==3.10
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.5
jedi==0.17.2
Jinja2==3.0.3
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.4
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
notebook==6.4.10
numpy==1.19.5
openai==0.8.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathlib==1.0.1
pdfminer.six==20211012
pdfplumber==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pvlib==0.9.1
pycparser==2.21
pydantic==1.9.2
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2023.3
pyzmq==25.0.2
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
terminado==0.13.0
testpath==0.6.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.20
uvicorn==0.16.0
Wand==0.6.13
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==3.6.4
zipp==3.6.0
